#50 refactor: position-store/intent-store の他 store 依存を _types/_constants 経由に変更
- _stores/_constants/index.ts を新設 (aggregation-index.md ルールに従い actor/actor-settings の constants.ts を re-export)
- position-store.ts・intent-store.ts が他 store の store.ts (実装) を直接 import せず、型 (_types) と定数 (_constants) のみに依存する形にする
