+++ src/prototypes/time-control/time-control-03/_stores/_constants/index.ts
@@ -1,0 +1,8 @@
+/**
+ * 各 store の constants.ts を再 export する集約 index
+ *
+ * - 定数分離済み (`_stores/<name>/constants.ts` を持つ) store が増えるたびに、\
+ *   ここに re-export を追加していく
+ */
+export * from '../actor-settings/constants'
+export * from '../actor/constants'
