+++ src/prototypes/time-control/time-control-03/_stores/intent-store.ts
@@ -4,8 +4,8 @@ import { getTickMs } from '../../time-control-02/_lib/get-tick-ms'
 import { generateMovePath } from '../_lib/generate-move-path'
 import { ActionLogEntry, MoveIntent } from '../types'
 
-import { ActorStore, DEFAULT_ACTOR_INFO } from './actor'
-import { ActorSettingsStore, DEFAULT_ACTOR_SETTINGS } from './actor-settings'
+import { DEFAULT_ACTOR_INFO, DEFAULT_ACTOR_SETTINGS } from './_constants'
+import { ActorSettingsStore, ActorStore } from './_types'
 import { GameClockStore } from './game-clock-store'
 import { PathStore } from './path-store'
 import { PlannedPathStore } from './planned-path-store'
